Implemented netplay
Added new CompressedInput struct for storing the input inside of a single int.
Rewrote Controller class to work with CompressedInput.
Added RawSerializeBuffer to write/read raw serialized data
Added netSerSave and netSerLoad functions to a bunch of classes, for serializing the state for rollback.
Added ShobuNetwork library (MIT license).
Rewrote part of the gameloop to work with rollback networking.