Implemented input mapper
Added InputMappingUI class to render a window where you can map the inputs of your controller.
Changed Gamepad to allow for setting new mapping in real time.
Changed Gamepad to allow buttons to be inversed.
Added dumpInput function to Gamepad which reads all possible input of a joystick into an array, this is used to identify buttons and axes in the input mapper.
Added getGamepadName function to Gamepad, which returns the name of the joystick.
Added debug functions to GamepadManager to get connected gamepads.